requirements.txt
openai
python-dotenv  # Eğer ortam değişkenlerini kullanacaksanız bu paketi ekleyebilirsiniz.

